File Associations!
- Rough initial body of work for allowing users to associate lyrics with any
other file.
- Creates the .lyrics file type and adds support for it in Lyricistant.

The idea behind this is that users can say "Hey, this lyrics are for this
instrumental file! It'd be neat if I could easily play that instrumental
while also writing the lyrics."

File associations support any file type, but if the file is an audio file that
we can play, we'll show a mini-player, allowing the user to play the referenced
file without leaving the app.